test(e2e): make sure FormData can be submitted
