+++ e2e/start/basic/app/routes/server-fns.tsx
@@ -5,6 +5,7 @@ import { ConsistentServerFnCalls } from './-server-fns/consistent-fn-calls'
 import { MultipartServerFnCall } from './-server-fns/multipart-formdata-fn-call'
 import { AllowServerFnReturnNull } from './-server-fns/allow-fn-return-null'
 import { SerializeFormDataFnCall } from './-server-fns/serialize-formdata-fn-call'
+import { SubmitFormDataFn } from './-server-fns/submit-formdata-fn'
 
 export const Route = createFileRoute('/server-fns')({
   component: RouteComponent,
@@ -17,6 +18,7 @@ function RouteComponent() {
       <MultipartServerFnCall />
       <AllowServerFnReturnNull />
       <SerializeFormDataFnCall />
+      <SubmitFormDataFn />
     </>
   )
 }
